Upgrade to Mycn 0.15

+++ sqlite-scripts/platform.sql
@@ -11,7 +11,7 @@ create table team (
 create table reg_team (
     reg_team_id integer not null primary key autoincrement,
     token varchar(255) not null unique,
-    team_id integer not null unique references team(team_id),
+    team_id bigint not null unique references team(team_id),
     user_email varchar(255) not null,
     user_name varchar(255) not null,
     user_login varchar(255) not null,
